requirements.txt
streamlit
openai
faiss-cpu
numpy
pandas
tqdm
python-dotenv
sentence-transformers
